build: Don't crash if the colophon is missing

+++ se/se_epub_build.py
@@ -136,14 +136,16 @@ def build(self, run_epubcheck: bool, build_kobo: bool, build_kindle: bool, outpu
 				file.truncate()
 
 			# Update the colophon with release info
-			with open(work_epub_root_directory / "epub" / "text" / "colophon.xhtml", "r+", encoding="utf-8") as file:
-				xhtml = file.read()
+			# If there's no colophon, skip this
+			if os.path.isfile(work_epub_root_directory / "epub" / "text" / "colophon.xhtml"):
+				with open(work_epub_root_directory / "epub" / "text" / "colophon.xhtml", "r+", encoding="utf-8") as file:
+					xhtml = file.read()
 
-				xhtml = xhtml.replace("<p>The first edition of this ebook was released on<br/>", f"<p>This edition was released on<br/>\n\t\t\t<b>{last_updated_friendly}</b><br/>\n\t\t\tand is based on<br/>\n\t\t\t<b>revision {self.last_commit.short_sha}</b>.<br/>\n\t\t\tThe first edition of this ebook was released on<br/>")
+					xhtml = xhtml.replace("<p>The first edition of this ebook was released on<br/>", f"<p>This edition was released on<br/>\n\t\t\t<b>{last_updated_friendly}</b><br/>\n\t\t\tand is based on<br/>\n\t\t\t<b>revision {self.last_commit.short_sha}</b>.<br/>\n\t\t\tThe first edition of this ebook was released on<br/>")
 
-				file.seek(0)
-				file.write(xhtml)
-				file.truncate()
+					file.seek(0)
+					file.write(xhtml)
+					file.truncate()
 
 		# Output the pure epub3 file
 		se.epub.write_epub(work_epub_root_directory, output_directory / epub_output_filename)
